Updated Advertisement logic

+++ src/main/java/kz/aqyl/newsdesk/service/impl/AdvertisementServiceImpl.java
@@ -2,9 +2,12 @@ package kz.aqyl.newsdesk.service.impl;
 
 import kz.aqyl.newsdesk.dto.AdvertisementDto;
 import kz.aqyl.newsdesk.entity.Advertisement;
+import kz.aqyl.newsdesk.exception.UnauthorizedException;
+import kz.aqyl.newsdesk.exception.UserAdvertisementNotFoundException;
 import kz.aqyl.newsdesk.mapper.AdvertisementMapper;
 import kz.aqyl.newsdesk.repository.AdvertisementRepository;
 import kz.aqyl.newsdesk.service.AdvertisementService;
+import kz.aqyl.newsdesk.service.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
@@ -19,14 +22,17 @@ public class AdvertisementServiceImpl implements AdvertisementService {
 
   private final AdvertisementRepository advertisementRepository;
 
+  private final UserService userService;
+
   @Override
   public AdvertisementDto addAdvertisement(AdvertisementDto advertisementDto) {
     Advertisement advertisement = AdvertisementMapper.toEntity(advertisementDto);
+    advertisement.setUser(userService.getCurrentSessionUser());
     return AdvertisementMapper.toDto(advertisementRepository.save(advertisement));
   }
 
   @Override
-  public List<AdvertisementDto> getAdvertisement() {
+  public List<AdvertisementDto> getAdvertisements() {
     List<Advertisement> advertisementList = advertisementRepository.findAll();
     return advertisementList.stream()
             .map(AdvertisementMapper::toDto)
@@ -42,6 +48,10 @@ public class AdvertisementServiceImpl implements AdvertisementService {
     }
 
     Advertisement advertisement = advertisementOptional.get();
+    if (!advertisement.getUser().getId().equals(userService.getCurrentSessionUser().getId())){
+      throw new UnauthorizedException("You are not authorized to change this account");
+    }
+
     advertisement.setTitle(advertisementDto.title());
     advertisement.setDescription(advertisementDto.description());
     advertisement.setCost(advertisementDto.cost());
@@ -51,6 +61,17 @@ public class AdvertisementServiceImpl implements AdvertisementService {
 
   @Override
   public void deleteAdvertisement(Long id) {
+    Optional<Advertisement> advertisementOptional = advertisementRepository.findById(id);
+
+    if (advertisementOptional.isEmpty()){
+      throw new UserAdvertisementNotFoundException("Advertisement not found");
+    }
+
+    Advertisement advertisement = advertisementOptional.get();
+    if (!advertisement.getUser().getId().equals(userService.getCurrentSessionUser().getId())){
+      throw new UnauthorizedException("You are not authorized to delete this account");
+    }
+
     advertisementRepository.deleteById(id);
   }
 }
